Random -> RandomArrayList
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@1505 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/net/l1j/server/model/DungeonRandom.java
@@ -106,7 +106,7 @@ public class DungeonRandom {
 		int _heading;
 
 		private NewDungeonRandom(int[] newX, int[] newY, short[] newMapId, int heading) {
-			for (byte i = 0; i < 5; i++) {
+			for (int i = 0; i < 5; i++) {
 				_newX[i] = newX[i];
 				_newY[i] = newY[i];
 				_newMapId[i] = newMapId[i];
