+++ src/net/l1j/server/model/L1MobGroupSpawn.java
@@ -67,7 +67,7 @@ public class L1MobGroupSpawn {
 			if (minion.isZero()) {
 				continue;
 			}
-			for (byte i = 0; i < minion.getCount(); i++) {
+			for (int i = 0; i < minion.getCount(); i++) {
 				mob = spawn(leader, minion.getId());
 				if (mob != null) {
 					mobGroupInfo.addMember(mob);
